Moving logic to move_picture_to_desk
