+++ server/ansa/__init__.py
@@ -1,10 +1,11 @@
 import logging
 import superdesk
 
-from superdesk.signals import item_publish, item_moved
+from superdesk.signals import item_publish
 from superdesk.metadata.item import PUBLISH_STATES
 
 from .constants import PRODUCTS_ID
+from .move_picture_to_desk import on_publish_from_personal_space
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -18,7 +19,7 @@ def on_update(resource, updates, original):
 
 def on_publish(updates, original):
     assign_product_to_pictures(updates, original)
-    move_associated_pictures_from_personal_space(updates, original)
+    on_publish_from_personal_space(updates, original)
 
 
 def assign_product_to_pictures(updates, original):
@@ -46,134 +47,6 @@ def assign_product_to_pictures(updates, original):
             original.setdefault("associations", {}).update({key: item})
 
 
-def move_associated_pictures_from_personal_space(updates, original):
-    """When publishing from personal space, move associated pictures to the target desk.
-
-    Pictures are moved to the stage defined in move_picture_stage vocabulary.
-    """
-    from .move_picture_to_desk import get_picture_stage_name, find_stage_by_name, is_desk_enabled
-
-    # only handle personal space (no desk on original)
-    if original and original.get("task", {}).get("desk"):
-        return
-
-    # get the target desk - read from request args since set_desk hasn't run yet
-    from flask import request
-
-    desk_id = updates.get("task", {}).get("desk")
-    if not desk_id and request and request.args:
-        desk_id = request.args.get("desk_id")
-    if not desk_id or not is_desk_enabled(desk_id):
-        return
-
-    stage_name = get_picture_stage_name()
-    if not stage_name:
-        return
-
-    stage = find_stage_by_name(desk_id, stage_name)
-    if not stage:
-        return
-
-    associations = updates.get("associations") or (original or {}).get("associations") or {}
-    archive_service = superdesk.get_resource_service("archive")
-
-    for key, assoc in associations.items():
-        if not assoc or assoc.get("type") != "picture":
-            continue
-        if assoc.get("state") in PUBLISH_STATES:
-            continue
-
-        picture_id = assoc.get("_id")
-        if not picture_id:
-            continue
-
-        picture = archive_service.find_one(req=None, _id=picture_id)
-        if not picture:
-            continue
-
-        # skip if already on this desk/stage
-        if str(picture.get("task", {}).get("desk") or "") == str(desk_id) and str(
-            picture.get("task", {}).get("stage") or ""
-        ) == str(stage["_id"]):
-            continue
-
-        logger.info(
-            'move_associated_pictures_from_personal_space: moving picture "%s" to desk "%s" stage "%s"',
-            picture_id,
-            desk_id,
-            stage_name,
-        )
-
-        updated_task = dict(picture.get("task") or {})
-        updated_task.update({"desk": desk_id, "stage": stage["_id"]})
-        archive_service.system_update(
-            picture_id,
-            {"task": updated_task},
-            picture,
-        )
-
-
-def on_item_moved(sender, item, original, **kwargs):
-    """When an article is sent from personal space, send associated pictures to the same desk.
-
-    Pictures are moved to the stage defined in move_picture_stage vocabulary,
-    if the desk is enabled in move_picture_desks vocabulary.
-    """
-    from .move_picture_to_desk import get_picture_stage_name, find_stage_by_name, is_desk_enabled
-
-    # only handle send from personal space (original had no desk)
-    if original.get("task", {}).get("desk"):
-        return
-
-    desk_id = item.get("task", {}).get("desk")
-    if not desk_id or not is_desk_enabled(desk_id):
-        return
-
-    stage_name = get_picture_stage_name()
-    if not stage_name:
-        return
-
-    stage = find_stage_by_name(desk_id, stage_name)
-    if not stage:
-        return
-
-    associations = item.get("associations") or {}
-    archive_service = superdesk.get_resource_service("archive")
-
-    for key, assoc in associations.items():
-        if not assoc or assoc.get("type") != "picture":
-            continue
-
-        picture_id = assoc.get("_id")
-        if not picture_id:
-            continue
-
-        picture = archive_service.find_one(req=None, _id=picture_id)
-        if not picture:
-            continue
-
-        # skip if already on this desk/stage
-        if str(picture.get("task", {}).get("desk") or "") == str(desk_id) and str(
-            picture.get("task", {}).get("stage") or ""
-        ) == str(stage["_id"]):
-            continue
-
-        logger.info(
-            'on_item_moved: sending picture "%s" to desk "%s" stage "%s"',
-            picture_id,
-            desk_id,
-            stage_name,
-        )
-
-        updated_task = dict(picture.get("task") or {})
-        updated_task.update({"desk": desk_id, "stage": stage["_id"]})
-        archive_service.system_update(
-            picture_id,
-            {"task": updated_task},
-            picture,
-        )
-
-
 def udpate_sign_off(sender, item, **kwargs):
     """When item is auto published add -RED to sign_off."""
     if item.get("auto_publish"):
@@ -190,7 +63,6 @@ def init_app(app):
 
     app.on_update += on_update
     item_publish.connect(udpate_sign_off)
-    item_moved.connect(on_item_moved)
 
     priority_to_profile_mapping = {}
     for key, val in app.config["PRIORITY_TO_PROFILE_MAPPING"].items():
